Jitter, tests, rename TryWithDeadline() to Try()
The previous version of Try(), based on failed attempts, was rarely used.
Try(), previously TryWithDeadline(), uses a time.Duration based soft
deadline, which is more intuitive.  Checking the number of attempts can
still be done using New() and WaitCalledCount().  This breaks the api,
but it's worth it.

Jitter has been introduced.  The wait time for each iteration is the sum
of half of the current fibonacci value plus a random value in the range
[0, fib_val/2].

A simple unit test has also been added.